Added container builder and organized code
