+++ ConsoleApp1/CinemaUI/User Interface/Selectable/TextList.cs
@@ -1,54 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using CinemaUI.Utility;
-using CinemaUI.Builder;
 
-namespace CinemaUI.Selectables
+namespace CinemaUI
 {
-    public class Selectable
-    {
-        public bool Selected { get; protected set; } = false;
-
-        public virtual void Unselect() => Selected = false;
-        public virtual void Select() => Selected = true;
-
-    }
-
-    public class SelectableText : Selectable
-    {
-        private Window _redirect { get; set; }
-        private Paragraph _paragraph { get; set; }
-        public string Text
-        {
-            get => _paragraph.Text;
-            set => _paragraph.Text = value;
-        }
-        public ConsoleColor Foreground { get; set; }
-        public ConsoleColor Background { get; set; }
-
-        public SelectableText(Paragraph paragraph, Color color)
-        {
-            this._paragraph = paragraph;
-            Foreground = color.Foreground;
-            Background = color.Background;
-        }
-
-        public override void Unselect()
-        {
-            Selected = false;
-            _paragraph.ChangeTextCells(_paragraph.TextColor, ConsoleColor.Black, true);
-        }
-        public override void Select()
-        {
-            Selected = true;
-            _paragraph.ChangeTextCells(Foreground, Background, Background != Foreground);
-        }
-    }
-
     public class SelectableList : Selectable
     {
-        private TextList TextList { get; set; }
+        internal TextList TextList { get; set; }
         internal List<SelectableText> Items { get; set; } = new List<SelectableText>();
         private int OrderIndex { get => TextList.Window.SelectionOrder.IndexOf(this); }
         protected int PriorityIndex { get; set; } = 0;
@@ -61,7 +19,6 @@ namespace CinemaUI.Selectables
             this.TextList = textList;
             Foreground = color.Foreground;
             Background = color.Background;
-            textList.Window.SelectionOrder.Add(this);
         }
 
         public void KeyResponse(ConsoleKeyInfo keyPressed)
@@ -91,6 +48,7 @@ namespace CinemaUI.Selectables
                     break;
             }
         }
+
         public override void Unselect()
         {
             Selected = false;
@@ -105,36 +63,8 @@ namespace CinemaUI.Selectables
     }
 }
 
-namespace CinemaUI.Selectables.Builder
+namespace CinemaUI.Builder
 {
-
-    public class SelectableTextBuilder : IBuilder
-    {
-        private SelectableText _product { get; set; }
-        private Tuple<Paragraph, Color> _params { get; set; }
-
-        public void Reset()
-        {
-            this._product = new SelectableText(_params.Item1, _params.Item2);
-        }
-
-        public SelectableTextBuilder(Paragraph paragraph, Color color)
-        {
-            this._params = new Tuple<Paragraph, Color>(paragraph, color);
-            this.Reset();
-        }
-
-        public SelectableText Result(string text)
-        {
-            SelectableText result = this._product;
-            result.Text = text;
-
-            this.Reset();
-
-            return result;
-        }
-    }
-
     public class SelectableGroupBuilder : IBuilder
     {
         private SelectableList _product { get; set; }
@@ -156,6 +86,10 @@ namespace CinemaUI.Selectables.Builder
         {
             SelectableList result = this._product;
             this.Reset();
+
+            result.TextList.Window.SelectionOrder.Add(result);
+            if (result.TextList.Window.SelectionOrder.IndexOf(result) == 0)
+                result.Select();
 
             return result;
         }
