+++ ConsoleApp1/Program.cs
@@ -1,24 +1,33 @@
 ﻿using System;
 using CinemaUI;
 using CinemaUI.Builder;
-using CinemaUI.Utility;
 
 namespace CinemaApplication
 {
     class Program
     {
+        static void MainScreen()
+        {
+            var w = new Window(true);
+            var list1 = new TextListBuilder(w, 2, 2).Selectable(
+                    textColor: ConsoleColor.White,
+                    selectionColor: new Color(ConsoleColor.White, ConsoleColor.DarkBlue),
+                    useNumbers: true,
+                     "Pizza", "Patat", "Pannekoek", "Soep", "Stampot"
+                    ).Result();
+            var list2 = new TextListBuilder(w, 20, 2).Selectable(
+                textColor: ConsoleColor.White,
+                selectionColor: new Color(ConsoleColor.Black, ConsoleColor.White),
+                useNumbers: true,
+                "Groente", "Fruit", "Vlees", "Vegetarisch", "Drinken"
+                ).Result();
+        }
+
         static void Main(string[] args)
         {
-            Console.ForegroundColor = ConsoleColor.White;
-            var w1 = new Window(true);
-            var testList = new TextListBuilder(w1, 5, 5)
-                .Selectable(ConsoleColor.White, new Color(ConsoleColor.White, ConsoleColor.DarkBlue), true,
-                "test1", "test2", "test3", "test4"
-                ).Result();
+            MainScreen();
+            
             InputHandler.WaitForInput();
-            Console.SetCursorPosition(0, 10);
-            Console.ForegroundColor = ConsoleColor.Black;
-            Console.BackgroundColor = ConsoleColor.Black;
         }
     }
 }
